Enable printing and recording of throughput + loss on a periodic basis
This commit adds the ability to report (i.e. display to stdout) the following information on a periodic basis
 * step number
 * throughput
 * total_loss
 * mlm_loss
 * nsp_loss
 * learning_rate

The frequency of the reporting is specified via the `--num_report_steps` option.

Currently only the `throughput` and `total_loss` values get recorded (to the trace events file meant for tensorboard consumption).

Note that `throughput` is the same as `examples.sec` and `total_loss` is the same as `loss` both of which are already reported and recorded via the `TPUEstimator` implementation.

The `LogSessionRunHook` class is based on a similar class in the NVBERT implementation. It can be easily enhanced to report and record other/more variables.